+++ pkg/sass_api/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.0-beta.14
 
-* No user-visible changes.
+* Fix a bug where `RecursiveAstVisitor.visitAtRootRule` wouldn't visit any nodes
+  interpolated into the `@at-root`'s query.
 
 ## 1.0.0-beta.13